move ColorScheme to types directory
