+++ src/types/ColorScheme.type.ts
@@ -1,0 +1,1 @@
+export type ColorScheme = 'dark' | 'light';
